Add timing

+++ insertion_serialization_anomaly/schema.sql
@@ -1,6 +1,6 @@
 DROP TABLE IF EXISTS "serializability_2";
 
 CREATE TABLE "serializability_2" (
-    "a" BIGINT PRIMARY KEY,
-    "b" INT NOT NULL
+    "a" INT PRIMARY KEY,
+    "b" BIGINT NOT NULL
 );
